ARC-12 Renamed data classes
