--- src/main/java/com/amtinez/api/rest/crud/datas/AuthorityData.java
@@ -1,14 +1,0 @@
-package com.amtinez.api.rest.crud.datas;
-
-import lombok.Data;
-
-/**
- * @author amartinezcerro@gmail.com
- */
-@Data
-public class AuthorityData {
-
-    private Long id;
-    private String authority;
-
-}
