+++ src/main/java/com/amtinez/api/rest/crud/dtos/Authority.java
@@ -1,0 +1,14 @@
+package com.amtinez.api.rest.crud.dtos;
+
+import lombok.Data;
+
+/**
+ * @author amartinezcerro@gmail.com
+ */
+@Data
+public class Authority {
+
+    private Long id;
+    private String name;
+
+}
