+++ src/main/java/com/amtinez/api/rest/crud/dtos/User.java
@@ -1,4 +1,4 @@
-package com.amtinez.api.rest.crud.datas;
+package com.amtinez.api.rest.crud.dtos;
 
 import lombok.Data;
 
@@ -9,13 +9,13 @@ import java.util.Set;
  * @author amartinezcerro@gmail.com
  */
 @Data
-public class UserData {
+public class User {
 
     private Long id;
     private String firstName;
     private String lastName;
     private String email;
     private LocalDateTime birthdayDate;
-    private Set<AuthorityData> authorities;
+    private Set<Authority> authorities;
 
 }
